Add initial Rust implementation of schema
Some initial workings at implementing the schema using Rust. I struggled
greatly with our schema and was able to wrap this within an `enum` but I
believe this is wrong.  The interactions with messages becomes much more
difficult since they require a wrap within a `match` statement in order
to get access to any of the fields within the message.

I named this sub-project, rust-prate, as in to prattle on.  I expect
this to change but wanted to have a fun name.