JsonObject -> LugChatMessage (client)
Replaced all reference to raw JsonObject with LugChatMessage in
the client (server still to come). Also added a nifty method
shortName(message, numChars) which produces a short name based on
the keyHash in the LCM with approximately the given number of
characters. Current implementation is to get the base64 keyHash,
and convert to a string of hex characters (truncated to fit the
given numChars).