Refactor into using async tasks via Tokio
Originally attempted to split the WebSocketStream into multiple threads
for dealing with sending and receiving separately.  Ran into many
problems and went with a single task that handles both.

I looked into passing closures between tasks and ran into numerous
issues.  I believe the best way to tackle this is to create another task
that listens to the channel and then reacts instead because passing a
closure that way should be easy.