Add disconnect and global panic logic
This is a "best attempt" implementation of sending a disconnect to the
server before shutdown.  Because of how the borrow controller in Rust
and Tauri works, the best implementation is to signal the socket thread.
The whole process should wait for the socket thread to complete but I
don't believe this is guaranteed.